mailbox: refactor Noise code for easy extensibility
Refactor the noise handshakestate code. This commit makes the code more
reusable and takes advantage of the fact that the message pattern token
writing and reading are completely separate from the payload writing and
reading. With this commit, we also have all the cryto code necessary for
when we add the `KK` pattern for the second handshake. A test is added
for the KK pattern here too.